Update get_ad_stats_columns.sql
Due to the lack of the alias tag and the column name not matching the source column, the AD_ID and AD_GROUP fields were casting as null in the stg_google_ads__ad_stats model when it tried to use the  fivetran_utils.fill_staging_columns macro. This fixes that issue and addresses Bug #31 in the Ad_Reporting package.